+++ picocli-examples/src/test/java/picocli/MapOptionsOptionalTest.java
@@ -8,6 +8,7 @@ import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TestRule;
 import picocli.CommandLine.Option;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -19,6 +20,8 @@ import static org.junit.Assert.*;
  */
 public class MapOptionsOptionalTest {
 
+    final static String ERRORMSG = CommandLine.Model.RuntimeTypeInfo.ERRORMSG;
+
     // allows tests to set any kind of properties they like, without having to individually roll them back
     @Rule
     public final TestRule restoreSystemProperties = new RestoreSystemProperties();
@@ -26,7 +29,6 @@ public class MapOptionsOptionalTest {
     @Rule
     public final ProvideSystemProperty ansiOFF = new ProvideSystemProperty("picocli.ansi", "false");
 
-    @Ignore("Requires #1108")
     @Test
     public void testOptionalIfNoValue() {
         class App {
@@ -34,10 +36,19 @@ public class MapOptionsOptionalTest {
         }
         App app = CommandLine.populateCommand(new App(), "-Dkey");
         assertEquals(1, app.map.size());
+        assertEquals(Optional.of(""), app.map.get("key"));
+    }
+
+    @Test
+    public void testOptionalEmptyIfNoValueWithFallbackNull() {
+        class App {
+            @Option(names = "-D", fallbackValue = "_NULL_") Map<String, Optional<String>> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey");
+        assertEquals(1, app.map.size());
         assertEquals(Optional.empty(), app.map.get("key"));
     }
 
-    @Ignore("Requires #1108")
     @Test
     public void testOptionalWithValue() {
         class App {
@@ -48,7 +59,6 @@ public class MapOptionsOptionalTest {
         assertEquals(Optional.of("value"), app.map.get("key"));
     }
 
-    @Ignore("Requires #1108")
     @Test
     public void testOptionalIfNoValueMultiple() {
         class App {
@@ -56,11 +66,21 @@ public class MapOptionsOptionalTest {
         }
         App app = CommandLine.populateCommand(new App(), "-Dkey1", "-Dkey2");
         assertEquals(2, app.map.size());
+        assertEquals(Optional.of(""), app.map.get("key1"));
+        assertEquals(Optional.of(""), app.map.get("key2"));
+    }
+
+    @Test
+    public void testOptionalIfNoValueMultipleWithFallbackNull() {
+        class App {
+            @Option(names = "-D", fallbackValue = "_NULL_") Map<String, Optional<String>> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey1", "-Dkey2");
+        assertEquals(2, app.map.size());
         assertEquals(Optional.empty(), app.map.get("key1"));
         assertEquals(Optional.empty(), app.map.get("key2"));
     }
 
-    @Ignore("Requires #1108")
     @Test
     public void testOptionalWithValueMultiple() {
         class App {
@@ -70,5 +90,79 @@ public class MapOptionsOptionalTest {
         assertEquals(2, app.map.size());
         assertEquals(Optional.of("val1"), app.map.get("key1"));
         assertEquals(Optional.of("val2"), app.map.get("key2"));
+    }
+    @Test
+    public void testBooleanIfNoValueMultiple() {
+        class App {
+            @Option(names = "-E", fallbackValue = "true") Map<String, Boolean> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Ekey1", "-Ekey2");
+        assertEquals(2, app.map.size());
+        assertEquals(Boolean.TRUE, app.map.get("key1"));
+        assertEquals(Boolean.TRUE, app.map.get("key2"));
+    }
+
+    @Test
+    public void testOptionalIntegerIfNoValueMultiple() {
+        class App {
+            @Option(names = "-D", fallbackValue = "_NULL_") Map<String, Optional<Integer>> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey1", "-Dkey2");
+        assertEquals(2, app.map.size());
+        assertEquals(Optional.empty(), app.map.get("key1"));
+        assertEquals(Optional.empty(), app.map.get("key2"));
+    }
+
+    @Test
+    public void testOptionalIntegerWithValueMultiple() {
+        class App {
+            @Option(names = "-D") Map<String, Optional<Integer>> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey1=123", "-Dkey2=456");
+        assertEquals(2, app.map.size());
+        assertEquals(Optional.of(123), app.map.get("key1"));
+        assertEquals(Optional.of(456), app.map.get("key2"));
+    }
+
+    @Test
+    public void testMapWithOptionalKeysNotSupported() {
+        class App {
+            @Option(names = "-D") Map<Optional<String>, Optional<Integer>> map;
+        }
+        try {
+            App app = CommandLine.populateCommand(new App(), "-Dkey1=123", "-Dkey2=456");
+            fail("Expected exception");
+        } catch (Exception ex) {
+            String msg = String.format(ERRORMSG, "java.util.Map<java.util.Optional<java.lang.String>, java.util.Optional<java.lang.Integer>>");
+            assertEquals(msg, ex.getMessage());
+        }
+    }
+
+    @Test
+    public void testListOfOptionalsNotSupported() {
+        class App {
+            @Option(names = "-X") List<Optional<Integer>> list;
+        }
+        try {
+            App app = CommandLine.populateCommand(new App(), "-X123", "-X456");
+            fail("Expected exception");
+        } catch (Exception ex) {
+            String msg = String.format(ERRORMSG, "java.util.List<java.util.Optional<java.lang.Integer>>");
+            assertEquals(msg, ex.getMessage());
+        }
+    }
+
+    @Test
+    public void testArrayOfOptionalsNotSupported() {
+        class App {
+            @Option(names = "-X") Optional<Integer>[] array;
+        }
+        try {
+            App app = CommandLine.populateCommand(new App(), "-X123", "-X456");
+            fail("Expected exception");
+        } catch (Exception ex) {
+            String msg = String.format(ERRORMSG, "java.util.Optional<java.lang.Integer>[]");
+            assertEquals(msg, ex.getMessage());
+        }
     }
 }